Taking a stab at initial project organization
