+++ helpers/error.go
@@ -1,0 +1,7 @@
+package helpers
+
+func CheckErr(err error) {
+	if err != nil {
+		panic(err) // Don't forget your towel
+	}
+}
